move hton* so that it can be used for non-pulpo initialized environment
